+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - It’s now possible to query for a single product or variant via GraphQL.
 - It’s now possible to access the Edit Customer page from the Edit Order page. ([#1397](https://github.com/craftcms/commerce/issues/1397))
+- It’s now possible to copy an address between address types on the Edit Order page. ([#1412](https://github.com/craftcms/commerce/issues/1412))
 - Address models now include `countryIso`. ([#1419](https://github.com/craftcms/commerce/issues/1419))
 - Added `craft\commerce\fields\Products::getContentGqlType()`.
 - Added `craft\commerce\fields\Variants::getContentGqlType()`.